requirements.txt
asgiref==3.2.7
boto3==1.12.33
botocore==1.15.33
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2.10
django-rosetta==0.9.3
django-storages==1.9.1
django-summernote==0.8.11.6
docutils==0.15.2
idna==2.9
jmespath==0.9.5
Pillow==7.1.0
polib==1.1.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
